refactor Pentagram component to handle form submission and remove SubmitButton
